requirements.txt
aiofiles~=24.1.0
annotated-types~=0.7.0
anyio~=4.8.0
async-timeout~=5.0.1
bcrypt~=4.2.1
certifi~=2025.1.31
charset-normalizer~=3.4.1
click~=8.1.8
coredis~=4.18.0
Deprecated~=1.2.18
dnspython~=2.7.0
email_validator~=2.2.0
faker~=33.0.0
fastapi~=0.115.11
geographiclib~=2.0
geopy~=2.4.1
h11~=0.14.0
httptools~=0.6.4
idna~=3.10
iniconfig~=2.0.0
mypy~=1.13.0
mypy-extensions~=1.0.0
fireo~=2.2.2
packaging~=24.2
passlib~=1.7.4
phonenumbers~=8.13.55
pluggy~=1.5.0
pydantic~=2.10.6
pydantic-extra-types~=2.10.2
pydantic-settings~=2.7.1
pydantic_core~=2.27.2
PyJWT~=2.9.0
Pympler~=1.1
pytest~=8.3.5
pytest-asyncio~=0.24.0
python-dateutil~=2.9.0.post0
python-dotenv~=1.0.1
python-multipart~=0.0.20
pytz~=2025.1
pyyaml~=6.0.2
requests~=2.32.3
six~=1.17.0
sniffio~=1.3.1
sse-starlette~=2.1.3
starlette~=0.41.3
typing_extensions~=4.12.2
urllib3~=2.3.0
uvicorn~=0.32.1
watchfiles~=1.0.4
websockets~=14.2
wrapt~=1.17.2
geopy~=2.4.1
httpx~=0.28.1